Split test_run.py into two files
